pyproject.toml
[project]
name = "livekit-outbound-agent"
version = "0.1.0"
description = "Production-ready outbound AI voice agent using LiveKit, Deepgram, Groq, and Cartesia"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[deepgram,cartesia,silero]>=0.12.0",
    "livekit-api>=0.6.0",
    "python-dotenv>=1.0.0",
    "groq>=0.5.0",
    "openai>=2.32.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
